'fixed' SLD generation to always create an int after scaling might not be what we really want

+++ src/main/java/org/openaviationmap/rendering/ScaleSLD.java
@@ -767,7 +767,7 @@ public final class ScaleSLD {
                         sb.append(' ');
                     } else {
                         double d = UOM.scaleValue(t, scale, dpi);
-                        sb.append(Double.toString(d));
+                        sb.append(Integer.toString((int) d));
                         sb.append(' ');
                     }
                 } catch (RenderException e) {
